fix(tilemap): avoid int overflow in resize() tile-count calculation
new_width * new_height is computed as int*int and only then promoted to
size_t for std::vector::resize(). For large maps (editors can produce
width/height whose product exceeds INT_MAX) the int multiplication
overflows before the promotion, yielding a too-small/wrong allocation.

Cast both operands to size_t so the multiplication happens in the wider
type, matching the resize(size_t) target.

+++ src/object/tilemap.cpp
@@ -656,7 +656,7 @@ TileMap::resize(int new_width, int new_height, int fill_id,
     }
   }
 
-  m_tiles.resize(new_width * new_height, fill_id);
+  m_tiles.resize(static_cast<size_t>(new_width) * static_cast<size_t>(new_height), fill_id);
 
   if (new_width > m_width) {
     // remap tiles
